Rename grpahml-base package
