+++ core/graphml-base/src/main/java/io/cml/graphml/base/Utils.java
@@ -12,7 +12,7 @@
  * limitations under the License.
  */
 
-package io.cml.graphml;
+package io.cml.graphml.base;
 
 import java.security.SecureRandom;
 import java.util.Arrays;
