+++ core/graphml-validation/src/main/java/io/cml/graphml/connector/canner/CannerClient.java
@@ -48,7 +48,7 @@ import static io.airlift.http.client.Request.Builder.preparePost;
 import static io.airlift.http.client.StaticBodyGenerator.createStaticBodyGenerator;
 import static io.airlift.http.client.StringResponseHandler.createStringResponseHandler;
 import static io.airlift.json.JsonCodec.jsonCodec;
-import static io.cml.graphml.Utils.randomIntString;
+import static io.cml.graphml.base.Utils.randomIntString;
 import static java.lang.String.format;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Objects.requireNonNull;
