+++ core/graphml-validation/src/main/java/io/cml/graphml/validation/ModelNameValidation.java
@@ -14,9 +14,9 @@
 
 package io.cml.graphml.validation;
 
-import io.cml.graphml.GraphML;
+import io.cml.graphml.base.GraphML;
+import io.cml.graphml.base.dto.Model;
 import io.cml.graphml.connector.Client;
-import io.cml.graphml.dto.Model;
 
 import java.time.Duration;
 import java.time.temporal.ChronoUnit;
