+++ core/graphml-validation/src/main/java/io/cml/graphml/validation/NotNullValidation.java
@@ -14,10 +14,10 @@
 
 package io.cml.graphml.validation;
 
-import io.cml.graphml.GraphML;
+import io.cml.graphml.base.GraphML;
+import io.cml.graphml.base.dto.Column;
 import io.cml.graphml.connector.AutoCloseableIterator;
 import io.cml.graphml.connector.Client;
-import io.cml.graphml.dto.Column;
 
 import java.time.Duration;
 import java.time.temporal.ChronoUnit;
